+++ flask_monitoringdashboard/frontend/js/app.js
@@ -41,7 +41,6 @@ import { DatabaseManagementController } from './controllers/databaseManagementCo
 import { EndpointVersionIPController } from './controllers/endpointVersionIP';
 import { EndpointVersionController } from "./controllers/endpointVersion";
 import { MonitorLevelController } from "./controllers/monitorLevel";
-import { SurveyController } from "./controllers/surveyController";
 import { TelemetryController } from "./controllers/telemetryController";
 
 
@@ -85,10 +84,6 @@ app.controller('EndpointController', ['$scope', 'endpointService', EndpointContr
 app.controller('PaginationController', ['$scope', 'paginationService', PaginationController]);
 app.controller('ModalController', ['$scope', '$window', '$browser', 'modalService', ModalController]);
 
-app.component('surveyComponent', {
-    templateUrl: 'static/pages/survey.html',
-    controller: SurveyController
-});
 app.component('telemetryComponent', {
     templateUrl: 'static/pages/telemetry.html',
     controller: TelemetryController