+++ flask_monitoringdashboard/views/telemetry.py
@@ -3,7 +3,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from flask_monitoringdashboard.core.auth import secure
 from flask_monitoringdashboard import blueprint, telemetry_config
 from flask_monitoringdashboard.core.config import TelemetryConfig
-from flask_monitoringdashboard.core.telemetry import get_telemetry_user
+from flask_monitoringdashboard.core.telemetry import get_telemetry_user, post_to_back_if_telemetry_enabled
 from flask_monitoringdashboard.database import session_scope
 
 
@@ -29,22 +29,6 @@ def accept_telemetry_consent():
     return '', 204
 
 
-@blueprint.route('/telemetry/survey_has_been_filled', methods=['GET'])
-def survey_has_been_filled():
-    with session_scope() as session:
-        try:
-            telemetry_user = get_telemetry_user(session)
-            telemetry_user.survey_filled = TelemetryConfig.ACCEPTED
-            session.commit()
-
-        except SQLAlchemyError as e:
-            print('error committing survey consent to database', e)
-            session.rollback()
-
-    # Return no content
-    return '', 204
-
-
 @blueprint.route('/telemetry/get_is_telemetry_answered', methods=['GET'])
 def get_is_telemetry_answered():
     with session_scope() as session:
@@ -52,10 +36,13 @@ def get_is_telemetry_answered():
         res = True if telemetry_user.monitoring_consent in (TelemetryConfig.REJECTED, TelemetryConfig.ACCEPTED) else False
         return {'is_telemetry_answered': res}
 
+@blueprint.route('/telemetry/submit_follow_up', methods=['POST'])
+def submit_follow_up():
+    data = request.json
+    feedback = data.get('feedback')
+    
+    post_to_back_if_telemetry_enabled('FollowUp', feedback=feedback)
 
-@blueprint.route('/telemetry/get_is_survey_filled', methods=['GET'])
-def get_is_survey_filled():
-    with session_scope() as session:
-        telemetry_user = get_telemetry_user(session)
-        res = True if telemetry_user.survey_filled in (TelemetryConfig.REJECTED, TelemetryConfig.ACCEPTED) else False
-        return {'is_survey_filled': res}
+    return jsonify({'message': 'Feedback submitted successfully'}), 200
+
+